herdme/requirements.txt
asgiref==3.3.4
attrs==20.3.0
Django==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
factory-boy==3.2.0
Faker==8.1.2
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
